requirement.txt
matplotlib==3.1.1
numpy==1.17.2
opencv-python==4.1.1.26
Pillow==6.2.0
protobuf==3.9.1
pycparser==2.19
pyparsing==2.4.2
python-dateutil==2.8.0
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
tensorboardX==1.8
torch==1.0.0
imageio==2.5.0
tqdm==4.36.1